@@ -1,1 +1,1 @@
-ruby-2.0.0-p247
+ruby-2.3.0
